+++ ElementX/Sources/FlowCoordinators/UserSessionFlowCoordinatorStateMachine.swift
@@ -114,7 +114,6 @@ class UserSessionFlowCoordinatorStateMachine {
         configure()
     }
 
-    // swiftlint:disable:next cyclomatic_complexity
     private func configure() {
         stateMachine.addRoutes(event: .start, transitions: [.initial => .roomList(selectedRoomID: nil)])
         stateMachine.addRoutes(event: .startWithMigration, transitions: [.initial => .migration])